api/requirements.txt
aniso8601==9.0.1
blis==0.7.7
catalogue==2.0.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cymem==2.0.6
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl
Flask==2.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
langcodes==3.3.0
MarkupSafe==2.1.1
murmurhash==1.0.7
nltk==3.7
numpy==1.22.4
packaging==21.3
pandas==1.4.2
pathy==0.6.1
pdfminer==20191125
preshed==3.0.6
pycodestyle==2.8.0
pycryptodome==3.14.1
pydantic==1.8.2
pyenchant==3.2.2
pygame==2.1.2
PyMuPDF==1.19.6
pyparsing==3.0.9
PyPDF2==2.0.0
python-dateutil==2.8.2
pytz==2022.1
regex==2022.6.2
requests==2.27.1
science-parse-api==1.0.1
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
smart-open==5.2.1
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
thinc==8.0.17
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.64.0
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
wasabi==0.9.1
Werkzeug==2.1.2
